+++ packages/pebble/src/tokenizer/utils/tokenFromKeyword.ts
@@ -118,6 +118,7 @@ export function tokenFromKeyword(text: string): Token {
             // if (text === "new") return Token.New;
             // if (text === "undefined") return Token.Null;
             if (text === "namespace") return Token.Namespace;
+            if (text === "nested") return Token.Nested;
 
             // replaced by "int"
             // if (text === "number") return Token.Number;
@@ -168,6 +169,7 @@ export function tokenFromKeyword(text: string): Token {
                 break;
             }
             if (text === "trace") return Token.Trace;
+            if (text === "top") return Token.Top;
             // if (text === "try") return Token.Try;
             // if (text === "throw") return Token.Throw;
             // if (text === "typeof") return Token.TypeOf;